Settings: drop black theme and refactor for dark night mode [3/3]
Most of this commit was supposed to be in our introduction of Fruity
Pebbles v2.0 (original commit by @Mazda--), however since we were still
using a black theme, we discarded many of the changes in that commit.
However, now, we are using many changes from that commit due to our
deprecation of the black theme. This commit is also a partial revert
of ce3e389e3eca.

Change-Id: Ic32f03072e2e9dbc7928b0bbf17dbe10fa313606